requirements.txt
python-dotenv~=1.0.1
langchain~=0.3.0
langchain-community~=0.3.0
langchain-core~=0.3.5
langchain-ollama~=0.2.0
langchain-openai~=0.2.0
langgraph~=0.2.27
pydantic~=2.9.2
neo4j~=5.24.0
pandas~=2.2.3